backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Orchids AI-powered website clone"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "playwright>=1.48.0",
    "beautifulsoup4>=4.12.3",
    "pydantic>=2.9.0",
    "uvicorn[standard]>=0.30.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
]
